+++ app/src/main/java/com/example/car_rental_project/model/UserEntity.kt
@@ -1,12 +1,14 @@
 package com.example.car_rental_project.model
 
+import java.math.BigInteger
+
 data class UserEntity(
     val userId : String? = "",
     val username : String? = "",
     val email : String? = "",
     val profilePicture : String? = "",
     val phoneNumber : String? = "",
-    val money : Int? = 0,
+    val money : Long? = 0,
     val premium : Boolean? = false,
 )
